gcAppEngine_flaskAPI/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.13.2
google-cloud-storage
fuzzywuzzy
datetime
pandas
numpy
xlrd
fsspec
requests
gunicorn
html5lib